test(react-langchain): fold the committed-ref case into the existing suite
it reuses `createMockStream` and `makeThreadListAdapter` instead of duplicating the scaffolding, and asserts the captured runtime identity survives the abandoned render so the dispatch is measured against the committed instance.
